requirements.txt
fastapi
uvicorn
flask
pymongo
motor
python-dotenv
pydantic
requests
